Implement partial methods in FoldingElementGenerator for better separation of concerns and improved readability
